Move UserInfo from management to meta/types
