+++ common/management/src/lib.rs
@@ -21,6 +21,5 @@ pub use cluster::ClusterApi;
 pub use cluster::ClusterMgr;
 pub use stage::StageMgr;
 pub use stage::StageMgrApi;
-pub use user::user_api::UserInfo;
 pub use user::user_api::UserMgrApi;
 pub use user::user_mgr::UserMgr;
