+++ query/src/datasources/database/system/users_table_test.rs
@@ -16,8 +16,8 @@ use std::sync::Arc;
 
 use common_base::tokio;
 use common_exception::Result;
-use common_management::UserInfo;
 use common_meta_types::AuthType;
+use common_meta_types::UserInfo;
 use common_meta_types::UserPrivilege;
 use common_meta_types::UserQuota;
 use futures::TryStreamExt;
