+++ query/src/interpreters/interpreter_user_create.rs
@@ -15,7 +15,7 @@
 use std::sync::Arc;
 
 use common_exception::Result;
-use common_management::UserInfo;
+use common_meta_types::UserInfo;
 use common_meta_types::UserPrivilege;
 use common_meta_types::UserQuota;
 use common_planners::CreateUserPlan;
