+++ query/src/users/user_mgr.rs
@@ -14,8 +14,8 @@
 
 use common_exception::ErrorCode;
 use common_exception::Result;
-use common_management::UserInfo;
 use common_meta_types::AuthType;
+use common_meta_types::UserInfo;
 use common_meta_types::UserPrivilege;
 use sha2::Digest;
